chore[model]: move purchase and sale calculator into model transactions
